API

Pandoc.Pandoc — Module

Pandoc

Pandoc wrapper to read JSON AST from pandoc

See https://hackage.haskell.org/package/pandoc-types-1.23/docs/Text-Pandoc-Definition.html

Pandoc.Attr — Type

Attr: identifier, classes, key-value pairs

Pandoc.BlockQuote — Type

Block quote (list of blocks)

Pandoc.BulletList — Type

Bullet list (list of items, each a list of blocks)

Pandoc.Caption — Type

The caption of a table or figure, with optional short caption.

Pandoc.Cell — Type

A table cell.

Pandoc.Cite — Type

Citation (list of inlines)

Pandoc.Code — Type

Inline code (literal)

Pandoc.CodeBlock — Type

Code block (literal) with attributes

Pandoc.ColSpan — Type

The number of columns occupied by a cell; the width of a cell.

Pandoc.ColSpec — Type

The specification for a single table column.

Pandoc.ColWidth — Type

The width of a table column, as a percentage of the text width.

Pandoc.Converter — Type

Converter(input = "# header level 1") -> String
Converter(input = "# header level 1", from="markdown", to="html") -> String

This is a Converter options struct. It supports all of pandoc's command line arguments.

You can use it like so:

run(Converter(; input = "# Header level 1"))

Pandoc.DefinitionList — Type

Definition list. Each list item is a pair consisting of a term (a list of inlines) and one or more definitions (each a list of blocks)

Pandoc.Div — Type

Generic block container with attributes

Pandoc.Emph — Type

Emphasized text (list of inlines)

Pandoc.Header — Type

Header - level (integer) and text (inlines)

Pandoc.HorizontalRule — Type

Horizontal rule

Pandoc.Image — Type

Image: alt text (list of inlines), target

Pandoc.LineBlock — Type

Multiple non-breaking lines

Pandoc.LineBreak — Type

Hard line break

Pandoc.Link — Type

Hyperlink: alt text (list of inlines), target

Pandoc.ListAttributes — Type

List attributes.

The first element of the triple is the start number of the list.

Pandoc.Math — Type

TeX math (literal)

Pandoc.Note — Type

Footnote or endnote

Pandoc.OrderedList — Type

Ordered list (attributes and a list of items, each a list of blocks)

Pandoc.Para — Type

Paragraph

Pandoc.Plain — Type

Plain text, not a paragraph

Pandoc.Quoted — Type

Quoted text (list of inlines)

Pandoc.RawBlock — Type

Raw block

Pandoc.RawInline — Type

Raw inline

Pandoc.RowHeadColumns — Type

The number of columns taken up by the row head of each row of a 'TableBody'. The row body takes up the remaining columns.

Pandoc.RowSpan — Type

The number of rows occupied by a cell; the height of a cell.

Pandoc.ShortCaption — Type

A short caption, for use in, for instance, lists of figures.

Pandoc.SmallCaps — Type

Small caps text (list of inlines)

Pandoc.SoftBreak — Type

Soft line break

Pandoc.Space — Type

Inter-word space

Pandoc.Span — Type

Generic inline container with attributes

Pandoc.Str — Type

Text (string)

Pandoc.Strikeout — Type

Strikeout text (list of inlines)

Pandoc.Strong — Type

Strongly emphasized text (list of inlines)

Pandoc.Subscript — Type

Subscripted text (list of inlines)

Pandoc.Superscript — Type

Superscripted text (list of inlines)

Pandoc.Table — Type

Table, with attributes, caption, optional short caption, column alignments and widths (required), table head, table bodies, and table foot

Pandoc.TableBody — Type

A body of a table, with an intermediate head, intermediate body, and the specified number of row header columns in the intermediate body.

Pandoc.TableFoot — Type

The foot of a table.

Pandoc.TableHead — Type

The head of a table.

Pandoc.Target — Type

Link target (URL, title).

Pandoc.Underline — Type

Underlined text (list of inlines)